+++ CHANGELOG.md
@@ -20,11 +20,16 @@ changes.
 
 - HeadLogic Outcome is now being trace on every protocol step transition.
 
+
 - **BREAKING** `hydra-cardano-api` changes:
   + Remove `Hydra.Cardano.Api.SlotNo` module.
   + Replace `fromConsensusPointHF` with `fromConsensusPointInMode` and
     `toConsensusPointHF` with `toConsensusPointInMode`.
   + Re-export new `AcquiringFailure` type from `cardano-api`.
+
+- **BREAKING** Addressed short-comings in `hydra-plutus` scripts:
+  + Check presence of state token (ST) and that it's consistent against datum.
+  + Moved check to reimburse commits to head validator.
 
 - Change the way the internal wallet initializes its state [#621](https://github.com/input-output-hk/hydra/pull/621)
   + The internal wallet does now always query ledger state and parameters at the tip.